requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-cloud-firestore==2.13.1
google-cloud-auth==2.23.4
google-cloud-secret-manager==2.18.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
openai==1.3.7
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0
